opusfile: Ensure file has the same final length when seeking
There appears to be an underlying gstreamer bug here.
While the frame size of opus frames should always be 20ms it is 15.3
for the first, and only the first, frame when starting playback.

This appears to have the effect of an extra packet being generated at
the end of the file.